# openapi.yaml
openapi: 3.1.0
info:
  title: Obiente Support Intake API
  version: 1.0.0
  description: Product-neutral private support intake. All successful report submissions require explicit user confirmation in the calling client.
  license:
    name: GNU Affero General Public License v3.0
    identifier: AGPL-3.0-only
servers:
  - url: https://support.obiente.org
security: []
paths:
...
  /api/v1/admin/login:
    post:
      summary: Sign in a support maintainer
      operationId: adminLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        "200":
          description: Authenticated maintainer session. The server also sets an HTTP-only session cookie.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminSession" }
        "401": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
...
  /api/v1/admin/logout:
    post:
      summary: Sign out the current maintainer
      operationId: adminLogout
      security: [{ AdminSessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      responses:
        "204": { description: Session revoked. }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    AdminSessionCookie:
      type: apiKey
      in: cookie
      name: obiente_admin_session
  parameters:
...
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string, minLength: 43, maxLength: 43 }
  responses:
    Problem:
      description: A bounded, user-safe error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    AdminSession:
      type: object
      required: [contractVersion, username, csrfToken, expiresAt]
      properties:
        contractVersion: { const: 1 }
        username: { type: string }
        csrfToken: { type: string, minLength: 43, maxLength: 43 }
        expiresAt: { type: string, format: date-time }
...
    Problem:
      type: object
      required: [contractVersion, code, message]
      properties:
        contractVersion: { const: 1 }
        code: { type: string }
        message: { type: string }